requirements.txt
openai
langchain
langchain_core
langgraph
python-dotenv
langchain-openai
langchain-community